Rename 27-Remove_Element to 27.Remove_Element
将给定数字列表中指定的字符去掉，并返回数字列表的长度（原地修改不允许重新划分空间）
Given an array nums and a value val, remove all instances of that value in-place and return the new length.
Do not allocate extra space for another array, you must do this by modifying the input array in-place with O(1) extra memory.
The order of elements can be changed. It doesn't matter what you leave beyond the new length.